Indicate that getRunningScript can return null
This tripped me up when using it -- I wasn't sure how to test whether a running script existed. Not sure whether the type should be updated to `RunningScript | null` or not so I left as is and just changed the comment, but let me know if I should update the type too!